move StorageNodeActorConfig to dedicated file
