+++ justin-core/src/multi-jvm/scala/justin/db/StorageNodeActorSpec.scala
@@ -2,24 +2,10 @@ package justin.db
 
 import akka.cluster.Cluster
 import akka.cluster.ClusterEvent.{CurrentClusterState, MemberUp}
-import akka.remote.testconductor.RoleName
-import akka.remote.testkit.{MultiNodeConfig, MultiNodeSpec}
+import akka.remote.testkit.MultiNodeSpec
 import akka.testkit.ImplicitSender
-import com.typesafe.config.ConfigFactory
 
 import scala.concurrent.duration._
-
-object StorageNodeActorConfig extends MultiNodeConfig {
-  val seed: RoleName  = role("StorageNode-seed")
-  val node1: RoleName = role("StorageNode-1")
-  val node2: RoleName = role("StorageNode-2")
-
-  commonConfig(ConfigFactory.parseString("""
-    akka.actor.provider="akka.cluster.ClusterActorRefProvider"
-    # don't use sigar for tests, native lib not in path
-    akka.cluster.metrics.collector-class = akka.cluster.JmxMetricsCollector"""
-  ))
-}
 
 class StorageNodeActorSpecMultiJvmNode1 extends StorageNodeActorSpec
 class StorageNodeActorSpecMultiJvmNode2 extends StorageNodeActorSpec
